+++ src/components/Card/card.js
@@ -5,9 +5,9 @@ import { BiBed } from "react-icons/bi";
 import { BiBath } from "react-icons/bi";
 import { BiArea } from "react-icons/bi";
 import { FaPaw } from "react-icons/fa";
-import { colors } from "../styles/colors";
-import { typography } from "../styles/typography";
-import { fonts } from "../styles/fonts";
+import { colors } from "../../styles/colors";
+import { typography } from "../../styles/typography";
+import { fonts } from "../../styles/fonts";
 import { FaRegEdit } from "react-icons/fa";
 import { AiOutlineCloseCircle } from "react-icons/ai";
 import { AiFillHeart } from "react-icons/ai";
